requirements.txt
flask==2.0.1
google-generativeai==0.3.1
python-dotenv==0.19.0
gunicorn==20.1.0
requests==2.31.0
Werkzeug==2.0.1
